Feat: added discord support

+++ BotFramework/Core/BotEventArgs.cs
@@ -4,7 +4,7 @@ namespace Tef.BotFramework.Core
 {
     public class BotEventArgs : EventArgs
     {
-        public BotEventArgs(string text, long groupId, int userSenderId, string username)
+        public BotEventArgs(string text, long groupId, long userSenderId, string username)
         {
             Text = text;
             GroupId = groupId;
@@ -23,7 +23,7 @@ namespace Tef.BotFramework.Core
 
         public string Text { get; }
         public long GroupId { get; }
-        public int UserSenderId { get; }
+        public long UserSenderId { get; }
         public string Username { get; }
     }
 }
